requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
python-pptx
beautifulsoup4
requests
lxml
python-dotenv